IGNITE-24260 Remove redundant tests
